Re-work net_poll to support IPD and SMS
Major re-work of net_poll framework to natively support IPD and SMS.
This handles multi-message IPD transfers and does SMS in a much better
way (with little change to users). At this point, the car module has
both SMS and IP MSG support and reports in its location and status to
the server every 1 minute (when one or more apps are connected) or 15
minutes (when no apps are connected).