Update vehicle_thinkcity.c
Removed car_chargelimit_soclimit = -1;  and car_chargelimit_rangelimit = -1;
They are already set in the ohms.c

car_tpem is not set when the module is inserted in the car after fw-flash. It's set first after charge plugged or key on.
It worked after moving car_stale_temps = 60; (in the poll1 section) before the variable car_tpem is set from cantata.